Reduce architecture index bandwidth for mobile targets
Constraint: Production meshes top out at 3,416 vertices.
Rejected: Keep UInt32 | wastes index memory and excludes older mobile GPUs.
Confidence: high
Scope-risk: narrow
Directive: Preserve UInt16 unless a generated mesh exceeds 65,535 vertices.
Tested: Unity EditMode AL.Tests.EditMode.Architecture (288/288)
Not-tested: Manual visual review

+++ unity/Assets/AL/Scripts/Editor/Architecture/CrownlandsStormwrightProductionModelBuilder.cs
@@ -443,7 +443,7 @@ namespace AL.Editor.Architecture
             var mesh = new Mesh
             {
                 name = name,
-                indexFormat = IndexFormat.UInt32
+                indexFormat = IndexFormat.UInt16
             };
             mesh.SetVertices(vertices);
             mesh.SetNormals(normals);
