+++ unity/Assets/AL/Scripts/Editor/Architecture/UmbralTownHallProductionModelBuilder.cs
@@ -472,7 +472,7 @@ namespace AL.Editor.Architecture
             var mesh = new Mesh
             {
                 name = name,
-                indexFormat = IndexFormat.UInt32
+                indexFormat = IndexFormat.UInt16
             };
             mesh.SetVertices(vertices);
             mesh.SetNormals(normals);
